file cleanup remove dev files
