--- PHASE2_STEP1_COMPLETE.md
@@ -1,189 +1,0 @@
-# Phase 2 Step 1: MMA Configuration - COMPLETE
-
-## What Was Done
-
-### ✅ Configured DPAS MMA Atom
-
-**File:** `src/sycl/kernels/gemm_sqrsum/device/gemm_sqrsum_types.hpp`
-
-Added proper XPU DPAS (Deep Learning Processing Acceleration System) configuration:
-
-```cpp
-// MMA Atom: XE DPAS operation
-using MMAOperation = XE_DPAS_TT<8, float, Element>;
-using MmaAtom = MMA_Atom<MMAOperation>;
-```
-
-**What this means:**
-- `XE_DPAS_TT` = Intel XPU's matrix multiply instruction
-- `<8, float, Element>` = 8x16x8 DPAS atom, float accumulator, Element input type
-- Works for FP16, BF16, and FP32 inputs (Element is template parameter)
-
-### ✅ Configured Subgroup Layout
-
-```cpp
-// Subgroup layout: 16x16 subgroups with column-major ordering
-using SubgroupLayout = Layout<Shape<_16, _16, _1>, Stride<_16, _1, _0>>;
-```
-
-**What this means:**
-- 16x16 = 256 subgroups total per workgroup
-- Good occupancy for XPU hardware
-- Stride pattern determines how subgroups map to tile
-
-### ✅ Created TiledMMA
-
-```cpp
-using TiledMma = typename TiledMMAHelper<
-    MmaAtom, 
-    Layout<TileShape>, 
-    SubgroupLayout>::TiledMMA;
-```
-
-**What this does:**
-- Tiles the 8x16x8 DPAS atoms across the 256x256x16 tile
-- Distributes work across 256 subgroups
-- Each subgroup gets a portion of the GEMM to compute
-
-### ✅ Added Required Includes
-
-```cpp
-#include <cute/atom/mma_atom.hpp>
-#include "cutlass/gemm/collective/mma_helper_xe.hpp"
-```
-
-These provide:
-- `MMA_Atom` template
-- `XE_DPAS_TT` operation definition
-- `TiledMMAHelper` for creating TiledMMA
-
-## How It Works
-
-### DPAS Atom Size: 8x16x8
-
-The fundamental matrix multiply unit on Intel XPU:
-- **M dimension**: 8 rows
-- **N dimension**: 16 columns  
-- **K dimension**: 8 elements to accumulate
-
-### Tiling to 256x256x16
-
-Our target tile size requires:
-- **M direction**: 256 / 8 = 32 DPAS atoms
-- **N direction**: 256 / 16 = 16 DPAS atoms
-- **K direction**: 16 / 8 = 2 DPAS atoms
-
-With 16x16 = 256 subgroups, each subgroup gets:
-- 32 / 16 = 2 atoms in M
-- 16 / 16 = 1 atom in N
-- Total: 2 DPAS operations per subgroup per K iteration
-
-## Verification
-
-To verify this compiled correctly:
-
-```bash
-# Check generated files
-ls build/src/sycl/gemm_sqrsum_kernel_*_256x256x16.cpp
-
-# Should see:
-# gemm_sqrsum_kernel_half_256x256x16.cpp
-# gemm_sqrsum_kernel_bf16_256x256x16.cpp  
-# gemm_sqrsum_kernel_float_256x256x16.cpp
-```
-
-## What's Next: Step 2
-
-Now that MMA is configured, the mainloop can actually instantiate:
-
-**Current state:**
-- Mainloop template exists
-- But it's not wired to a complete kernel yet
-
-**Next step:**
-- Update mainloop to use real TiledMMA (remove placeholders)
-- Verify fragment types match
-- Ensure copy operations work with configured MMA
-
-See `GEMM_SQRSUM_PHASE2_TODO.md` Step 2 for details.
-
-## Technical Deep Dive
-
-### Why These Specific Parameters?
-
-**Why 8x16x8 DPAS?**
-- Hardware instruction size on Intel XPU
-- Optimal for FP16/BF16 matrix operations
-- Balances compute vs register pressure
-
-**Why 16x16 subgroup layout?**
-- 256 subgroups = good GPU occupancy
-- Square layout (16x16) balances M and N dimensions
-- Matches tile shape aspect ratio (256x256)
-
-**Why 256x256x16 tile?**
-- Large enough to amortize memory latency
-- Small enough to fit in cache/registers
-- K=16 chosen for smaller register footprint (square sum accumulation needs extra registers)
-
-### Comparison with MLA Kernel
-
-MLA decode kernel uses similar pattern:
-
-```cpp
-// From mla_decode_types.hpp
-using MMAOperation = XE_DPAS_TT<cute::gcd(SGTileQ, 8), float, ElementQ>;
-```
-
-Key difference:
-- MLA uses dynamic tile size based on `SGTileQ`
-- We use fixed 8 (standard DPAS size)
-- Both accumulate to float for numerical stability
-
-### Register Usage Estimate
-
-Each subgroup handles:
-- **A fragment**: 2 atoms × 8 rows × 8 k-elements = 128 elements
-- **B fragment**: 1 atom × 16 cols × 8 k-elements = 128 elements
-- **C fragment**: 2 × 1 × (8×16) = 256 elements (accumulator)
-- **SqrSum fragment**: 8 rows (one per row of A)
-
-Total per subgroup: ~128 + 128 + 256 + 8 = 520 FP16/BF16 elements
-In FP32 accumulators: ~256 + 8 = 264 FP32 elements
-
-This should fit comfortably in XPU registers (256 registers × 16 elements = 4096 FP32 capacity per subgroup).
-
-## Build Status
-
-Currently building - check output for:
-- ✅ No template instantiation errors
-- ✅ MMA_Atom types resolve correctly
-- ✅ TiledMMAHelper succeeds
-
-If errors occur, common issues:
-1. Missing includes for MMA operations
-2. Incompatible element types
-3. Tile size doesn't divide evenly by MMA atom size
-
-## Files Modified
-
-1. `src/sycl/kernels/gemm_sqrsum/device/gemm_sqrsum_types.hpp`
-   - Lines 64-88: Added MMA configuration
-   - Lines 16-18: Added includes
-
-## Next Session Plan
-
-1. **Wait for build to complete**
-2. **Check for compilation errors**
-3. **If successful**: Move to Step 2 (update mainloop)
-4. **If errors**: Debug MMA configuration issues
-
-## Estimated Remaining Work
-
-- Step 2 (Mainloop): 30 mins - mostly removing placeholders
-- Step 3 (Epilogue): 3-4 hours - the hard part
-- Step 4 (Kernel wrapper): 1-2 hours
-- Step 5 (Integration): 2-3 hours
-- Testing: 2-4 hours
-
-**Total remaining**: ~9-14 hours
