--- export_proxies.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-export http_proxy=http://proxy-chain.intel.com:911/
-export https_proxy=http://proxy-chain.intel.com:912/
-export no_proxy=127.0.0.1,localhost,.*.intel.com
-export HTTP_PROXY=http://proxy-chain.intel.com:911/
-export HTTPS_PROXY=http://proxy-chain.intel.com:912/
-export NO_PROXY=127.0.0.1,localhost,.*.intel.com
