+++ python/sgl_kernel/mhc.py
@@ -149,11 +149,28 @@ def hc_pre_big_fuse(
     )
 
 
-def hc_pre_gemm(A: torch.Tensor, B: torch.Tensor, C: torch.Tensor):
+def gemm_sqrsum(
+    C: torch.Tensor,
+    sqrsum: torch.Tensor,
+    A: torch.Tensor,
+    B: torch.Tensor,
+) -> None:
     """
-    HC Pre GEMM: C = A @ B
-    A: [T, K] bfloat16, row-major
-    B: [K, N] float32, column-major
-    C: [T, N] float32, row-major (output)
+    Compute C[m,n] = sum_k A[m,k]*B[n,k] and sqrsum[m] = sum_k A[m,k]^2.
+
+    This is the mhc_pre GEMM+sqrsum stage: B is the weight matrix fn given as
+    [N, K] (e.g. [24, 16384]), so C = A @ B^T. Leading singleton (n_splits=1)
+    axes on any argument are squeezed away.
+
+    Precision: when B is fp32 the kernel runs a tf32 x tf32 -> fp32 DPAS path
+    (A widened to fp32, B taken as-is, both reinterpreted to tf32 at load). When
+    A and B share a 16-bit dtype (half/bf16) the native DPAS path runs. C and
+    sqrsum are always fp32.
+
+    Args:
+        C: Output tensor [M, N] fp32, filled with A @ B^T
+        sqrsum: Output tensor [M] fp32, filled with row-wise squared sums of A
+        A: Input tensor [M, K]   (bf16/fp16/fp32)
+        B: Input tensor [N, K]   (fp32 for the production tf32 path)
     """
-    torch.ops.sgl_kernel.hc_pre_gemm.default(A, B, C)
+    torch.ops.sgl_kernel.gemm_sqrsum.default(C, sqrsum, A, B)
